refactor: text input cleanup and better seperation of files for adornment functionality
