+++ packages/text-input/src/TextInput.tsx
@@ -10,8 +10,8 @@ import { buildDataAttributes } from '@spark-web/utils/internal';
 import type { AllHTMLAttributes } from 'react';
 import { forwardRef } from 'react';
 
-import type { AdornmentsAsChildren } from './InputAdornment';
-import { childrenToAdornments } from './InputAdornment';
+import type { AdornmentsAsChildren } from './childrenToAdornments';
+import { childrenToAdornments } from './childrenToAdornments';
 
 type validTypes =
   | 'text'
