+++ packages/text-input/src/index.ts
@@ -3,5 +3,5 @@ export { TextInput, useInput } from './TextInput';
 
 // types
 
-export type { AdornmentChild } from './InputAdornment';
+export type { AdornmentChild } from './childrenToAdornments';
 export type { TextInputProps, UseInputProps } from './TextInput';
